webapi/tct-alarm-tizen-tests/alarm/AlarmManager_getAll_extra_argument.html

<!DOCTYPE html>

<!--

Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.



Authors:
        Mariusz Polasinski <m.polasinski@samsung.com>


-->

<html>

<head>
<title>AlarmManager_getAll_extra_argument</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
<script type="text/javascript" src="support/alarm_common.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: AlarmManager_getAll_extra_argument
//==== LABEL Check using getAll withe extra argument to get all alarms
//==== SPEC Tizen Web API:Application:Alarm:AlarmManager:getAll M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/alarm.html
//==== PRIORITY P1
//==== TEST_CRITERIA MNAEX MR

test(function () {
    add_result_callback(function () {
        try {
            tizen.alarm.removeAll();
        } catch (err) {
            // do nothing in case removeAll throw an exception
        }
    });

    var i, absAlarm, alarmAll = [], argumentsList = [null, undefined, "string", 1, false, ["one", "two"], {arg: 1}, function () {}];
    cleanAlarms();
    absAlarm = createAbsAlarm();
    tizen.alarm.add(absAlarm, APPLICATION_ID);
    for (i = 0; i < argumentsList.length; i++) {
        alarmAll = tizen.alarm.getAll(argumentsList[i]);
        assert_equals(alarmAll.length, 1, "number of added alarms should be 1");
        assert_true(alarmAll[0] instanceof tizen.AlarmAbsolute, "object was not get properly");
        assert_equals(alarmAll[0].id, absAlarm.id, "the first object is not the same which was added");
        alarmAll = [];
    }
}, document.title);

</script>
</body>
</html>
